+++ CMakeLists.txt
@@ -74,6 +74,10 @@ pybind11_add_module(qilisdk_libs SHARED)
 endif()
 file(GLOB LIBS_SOURCES src/qilisdk_cpp/libs/*.cpp)
 target_sources(qilisdk_libs PRIVATE ${LIBS_SOURCES})
+# Needed for dlopen/dlsym in the optional GPU shim (libs/cuda_solver.cpp).
+# No CUDA toolkit link dependency: cuBLAS/cuSOLVER are loaded lazily at runtime.
+# Empty on Windows (CMAKE_DL_LIBS), where the shim is compiled out.
+target_link_libraries(qilisdk_libs PRIVATE ${CMAKE_DL_LIBS})
 
 # Origin is Linux specific, but we can set it to @loader_path on macOS which has the same effect
 if(APPLE)
@@ -89,9 +93,6 @@ file(GLOB QILISIM_SUBDIR_SOURCES src/qilisdk_cpp/backends/qilisim/*/*.cpp)
 list(APPEND QILISIM_SOURCES ${QILISIM_SUBDIR_SOURCES})
 target_sources(qilisim_module PRIVATE ${QILISIM_SOURCES})
 target_link_libraries(qilisim_module PRIVATE qilisdk_libs)
-# Needed for dlopen/dlsym in the optional GPU shim (gpu/cuda_solver.cpp).
-# No CUDA toolkit link dependency: cuBLAS/cuSOLVER are loaded lazily at runtime.
-target_link_libraries(qilisim_module PRIVATE ${CMAKE_DL_LIBS})
 set_target_properties(qilisim_module PROPERTIES
     BUILD_RPATH "${RPATH_ORIGIN}"
     INSTALL_RPATH "${RPATH_ORIGIN}"
